refactor(cli): 按域拆分 TsBridgeRuntime 并归位测试
- runtime/ 下 config/session/mcp/extensions 等 impl 块
- 测试迁入 tests/ 子模块，集成测保留 Node bridge 依赖
- mod.rs 仅作门面（约 77 行）
