+++ apps/cli/src/ts_bridge/runtime/mod.rs
@@ -1,1 +1,10 @@
+pub(crate) mod approval;
+pub(crate) mod config;
+pub(crate) mod export;
+pub(crate) mod extensions;
 pub(crate) mod lifecycle;
+pub(crate) mod mcp;
+pub(crate) mod metadata;
+pub(crate) mod rewind;
+pub(crate) mod session;
+pub(crate) mod subagent;
